+++ src/layer/power.cpp
@@ -47,7 +47,7 @@ int Power::forward_inplace(Mat& bottom_top_blob, const Option& opt) const
 
         for (int i = 0; i < size; i++)
         {
-            ptr[i] = static_cast<float>(pow((shift + ptr[i] * scale), power));
+            ptr[i] = powf((shift + ptr[i] * scale), power);
         }
     }
